Code refactoring and optimization
