+++ src/Adapter/Shop/CommandHandler/UploadLogosHandler.php
@@ -29,7 +29,7 @@ namespace PrestaShop\PrestaShop\Adapter\Shop\CommandHandler;
 use PrestaShop\PrestaShop\Core\ConfigurationInterface;
 use PrestaShop\PrestaShop\Core\Domain\Shop\Command\UploadLogosCommand;
 use PrestaShop\PrestaShop\Core\Domain\Shop\CommandHandler\UploadLogosHandlerInterface;
-use PrestaShop\PrestaShop\Core\Form\ValueObject\ShopRestriction;
+use PrestaShop\PrestaShop\Core\Form\DTO\ShopRestriction;
 use PrestaShop\PrestaShop\Core\Hook\HookDispatcherInterface;
 use PrestaShop\PrestaShop\Core\Shop\LogoUploader;
 use Shop;
