+++ src/PrestaShopBundle/Controller/Admin/Improve/Design/ThemeController.php
@@ -43,8 +43,8 @@ use PrestaShop\PrestaShop\Core\Domain\Theme\Exception\CannotEnableThemeException
 use PrestaShop\PrestaShop\Core\Domain\Theme\Exception\ImportedThemeAlreadyExistsException;
 use PrestaShop\PrestaShop\Core\Domain\Theme\Exception\ThemeConstraintException;
 use PrestaShop\PrestaShop\Core\Domain\Theme\Exception\ThemeException;
-use PrestaShop\PrestaShop\Core\Domain\Theme\Query\GetThemeLogos;
-use PrestaShop\PrestaShop\Core\Domain\Theme\QueryResult\Logo;
+use PrestaShop\PrestaShop\Core\Domain\Shop\Query\GetLogosPaths;
+use PrestaShop\PrestaShop\Core\Domain\Shop\QueryResult\LogosPaths;
 use PrestaShop\PrestaShop\Core\Domain\Theme\ValueObject\ThemeImportSource;
 use PrestaShop\PrestaShop\Core\Domain\Theme\ValueObject\ThemeName;
 use PrestaShop\PrestaShop\Core\Form\FormHandlerInterface;
@@ -95,8 +95,8 @@ class ThemeController extends AbstractAdminController
 
         $themeProvider = $this->get('prestashop.core.addon.theme.theme_provider');
         $installedRtlLanguageChecker = $this->get('prestashop.adapter.language.rtl.installed_language_checker');
-        /** @var Logo $logoProvider */
-        $logoProvider = $this->getQueryBus()->handle(new GetThemeLogos());
+        /** @var LogosPaths $logoProvider */
+        $logoProvider = $this->getQueryBus()->handle(new GetLogosPaths());
 
         return $this->render('@PrestaShop/Admin/Improve/Design/Theme/index.html.twig', [
             'themeCatalogUrl' => $themeCatalogUrl,
