+++ src/PrestaShopBundle/Form/Admin/Type/ShopRestrictionCheckboxType.php
@@ -26,7 +26,6 @@
 
 namespace PrestaShopBundle\Form\Admin\Type;
 
-
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\CheckboxType;
 use Symfony\Component\Form\FormBuilderInterface;
@@ -36,6 +35,9 @@ use Symfony\Component\Form\FormInterface;
 use Symfony\Component\Form\FormView;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
+/**
+ * {@inheritdoc}
+ */
 class ShopRestrictionCheckboxType extends AbstractType
 {
     /**
