+++ src/PrestaShopBundle/Resources/config/services/adapter/shop.yml
@@ -15,3 +15,12 @@ services:
     class: 'PrestaShop\PrestaShop\Adapter\Shop\Url\BaseUrlProvider'
     arguments:
       - "@=service('prestashop.adapter.legacy.context').getContext().link"
+
+  prestashop.adapter.shop.query_handler.get_logos_paths_handler:
+    class: 'PrestaShop\PrestaShop\Adapter\Shop\QueryHandler\GetLogosPathsHandler'
+    arguments:
+      - '@=service("prestashop.adapter.legacy.configuration").get("_PS_IMG_")'
+      - '@=service("prestashop.adapter.legacy.configuration").get("_PS_IMG_DIR_")'
+    tags:
+      - name: 'tactician.handler'
+        command: PrestaShop\PrestaShop\Core\Domain\Shop\Query\GetLogosPaths
