+++ src/PrestaShopBundle/Resources/views/Admin/Improve/Design/Theme/Blocks/rtl_configuration.html.twig
@@ -37,7 +37,7 @@
           {{ 'Be careful! Please check your theme in an RTL language before generating the RTL stylesheet: your theme could be already adapted to RTL.\nOnce you click on "Adapt to RTL", any RTL-specific file that you might have added to your theme might be deleted by the created stylesheet.'|trans({}, 'Admin.Design.Help') }}
         </p>
       </div>
-      
+
       {{ ps.form_group_row(adaptThemeToRtlLanguagesForm.theme_to_adapt, {}, {
         'label': 'Theme to adapt'|trans({}, 'Admin.Design.Feature')
       }) }}
@@ -52,7 +52,7 @@
     <hr>
     <div class="card-block row">
       <div class="card-text">
-        {{ ps.multistore_switch(shopLogosForm.options, {
+        {{ ps.multistore_switch(shopLogosForm.shop_logos, {
           'label': 'Multistore'|trans({}, 'Admin.Global'),
         }) }}
       </div>
