+++ src/PrestaShopBundle/Resources/views/Admin/Improve/Design/Theme/index.html.twig
@@ -1,27 +1,27 @@
 {#**
-  * 2007-2018 PrestaShop
-  *
-  * NOTICE OF LICENSE
-  *
-  * This source file is subject to the Open Software License (OSL 3.0)
-  * that is bundled with this package in the file LICENSE.txt.
-  * It is also available through the world-wide-web at this URL:
-  * https://opensource.org/licenses/OSL-3.0
-  * If you did not receive a copy of the license and are unable to
-  * obtain it through the world-wide-web, please send an email
-  * to license@prestashop.com so we can send you a copy immediately.
-  *
-  * DISCLAIMER
-  *
-  * Do not edit or add to this file if you wish to upgrade PrestaShop to newer
-  * versions in the future. If you wish to customize PrestaShop for your
-  * needs please refer to http://www.prestashop.com for more information.
-  *
-  * @author    PrestaShop SA <contact@prestashop.com>
-  * @copyright 2007-2018 PrestaShop SA
-  * @license   https://opensource.org/licenses/OSL-3.0 Open Software License (OSL 3.0)
-  * International Registered Trademark & Property of PrestaShop SA
-  *#}
+ * 2007-2019 PrestaShop and Contributors
+ *
+ * NOTICE OF LICENSE
+ *
+ * This source file is subject to the Open Software License (OSL 3.0)
+ * that is bundled with this package in the file LICENSE.txt.
+ * It is also available through the world-wide-web at this URL:
+ * https://opensource.org/licenses/OSL-3.0
+ * If you did not receive a copy of the license and are unable to
+ * obtain it through the world-wide-web, please send an email
+ * to license@prestashop.com so we can send you a copy immediately.
+ *
+ * DISCLAIMER
+ *
+ * Do not edit or add to this file if you wish to upgrade PrestaShop to newer
+ * versions in the future. If you wish to customize PrestaShop for your
+ * needs please refer to https://www.prestashop.com for more information.
+ *
+ * @author    PrestaShop SA <contact@prestashop.com>
+ * @copyright 2007-2019 PrestaShop SA and Contributors
+ * @license   https://opensource.org/licenses/OSL-3.0 Open Software License (OSL 3.0)
+ * International Registered Trademark & Property of PrestaShop SA
+ *#}
 
 {% extends '@PrestaShop/Admin/layout.html.twig' %}
 
@@ -56,7 +56,7 @@
             {% include '@PrestaShop/Admin/Improve/Design/Theme/Blocks/rtl_configuration.html.twig' %}
         {% endif %}
 
-        {{ form_start(shopLogosForm.options, {'action': path('admin_themes_upload_logos')}) }}
+        {{ form_start(shopLogosForm.shop_logos, {'action': path('admin_themes_upload_logos')}) }}
           {% if not isInstalledRtlLanguage and isSingleShopContext and isMultiShopFeatureUsed %}
             {% include '@PrestaShop/Admin/Improve/Design/Theme/Blocks/multishop_switch.html.twig' %}
           {% endif %}
@@ -74,8 +74,8 @@
               <div class="clearfix">&nbsp;</div>
             </div>
           </div>
-        {{ form_rest(shopLogosForm.options) }}
-        {{ form_end(shopLogosForm.options) }}
+        {{ form_rest(shopLogosForm.shop_logos) }}
+        {{ form_end(shopLogosForm.shop_logos) }}
 
         <div class="card">
           <div class="card-header">
